Fixing issue #35 - SurfaceObserver not starting
Adds the Start() method back to SpatialMappingManager, which was
accidentally removed during the last update. This method is repsonsible
for starting the SurfaceObserver.
Also fixed: the RecalculateNormals was no longer present in
SpatialMappingObserver.cs. This call is required when using shaders
(like 'BlueLinesOnWalls') that rely on surface mesh normals being
correct.